Add resendRequest

+++ fds/src/main/java/com/distribox/fds/controllers/FilesController.java
@@ -6,11 +6,17 @@ import com.distribox.fds.entities.User;
 import com.distribox.fds.repos.FilesRepository;
 import com.distribox.fds.repos.ServersRepository;
 import com.distribox.fds.repos.UsersRepository;
+import com.distribox.fds.zookeeper.ZookeeperConfig;
+import org.apache.coyote.Request;
+import org.apache.curator.framework.recipes.leader.Participant;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriUtils;
 
 import java.nio.charset.StandardCharsets;
@@ -27,6 +33,12 @@ public class FilesController {
 
 	@Autowired
 	FilesRepository filesRepository;
+
+	@Autowired
+	ZookeeperConfig zookeeperConfig;
+
+//	@Autowired
+//	RestTemplate restTemplate;
 
 	Map<String, File> fileMap = new HashMap<>();
 
@@ -52,19 +64,20 @@ public class FilesController {
 			return ResponseEntity.badRequest().build();
 		}
 		User user;
-		Optional<File> existingFile = filesRepository.findByFilepath(filepath);
-		if (fileMap.containsKey(filepath)) {
-			return mergeServers(fileMap.get(filepath), serverids);
-		}
-		if (existingFile.isPresent()) {
-			return mergeServers(existingFile.get(), serverids);
-		}
 		if (!usersRepository.existsByUseridLike(userid)) {
 			user = new User(userid);
 			usersRepository.save(user);
 		} else {
 			user = usersRepository.findByUserid(userid);
 		}
+		Optional<File> existingFile = filesRepository.findById(filepath);
+		if (fileMap.containsKey(filepath)) {
+			return mergeServers(fileMap.get(filepath), serverids, body);
+		}
+		if (existingFile.isPresent()) {
+			return mergeServers(existingFile.get(), serverids, body);
+		}
+
 		Set<Server> serverSet = new HashSet<>();
 		for (String serverid : serverids) {
 			Server s = serversRepository.findById(serverid).get();
@@ -74,14 +87,17 @@ public class FilesController {
 		newFile.addServers(serverSet);
 		fileMap.put(newFile.getFilepath(), newFile);
 //		ResponseEntity<File> response = saveFileToDB(newFile);
+		System.out.println("hey!!!!");
+		resendRequest(HttpMethod.POST, "/saveFile", body);
 		return ResponseEntity.ok(newFile);
 	}
 
-	public ResponseEntity<File> mergeServers(File existingFile, List<String>newIds) {
+	public ResponseEntity<File> mergeServers(File existingFile, List<String>newIds, Map<String, Object> body) {
 		Set<Server>incomingServerSet = newIds.stream()
 				.map(id -> serversRepository.findById(id).get()).collect(Collectors.toSet());
 		existingFile.addServers(incomingServerSet);
 		fileMap.put(existingFile.getFilepath(), existingFile);
+		resendRequest(HttpMethod.POST, "/saveFile", body);
 		return ResponseEntity.ok(existingFile);
 	}
 
@@ -98,12 +114,24 @@ public class FilesController {
 	}
 
 	@PostMapping("/deleteFile")
-	public File deleteFilesReuqest(String filepath) {
-		File file = filesRepository.findByFilepath(filepath).get();
+	public ResponseEntity<File> deleteFilesReuqest(@RequestBody String filepath) {
+		String decodedPath = UriUtils.decode(filepath, StandardCharsets.UTF_8)
+				.replaceAll("filepath=", "");
+		System.out.println("filePath to delete: " + decodedPath);
+		Optional<File> fileOpt = filesRepository.findById(decodedPath);
+		if (fileOpt.isEmpty()) {
+			return ResponseEntity.notFound().build();
+		}
+		File file = fileOpt.get();
+		for (Server server: file.getServers()) {
+			file.removeServer(server);
+			serversRepository.save(server);
+		}
 		filesRepository.delete(file);
-		fileMap.remove(filepath);
+		fileMap.remove(decodedPath);
 		//TODO: Test that it's also gone from owning servers
-		return file;
+		resendRequest(HttpMethod.POST, "/deleteFile", filepath);
+		return ResponseEntity.ok(file);
 	}
 
 	/**
@@ -120,6 +148,7 @@ public class FilesController {
 				return ResponseEntity.badRequest().body("Invalid filepath");
 		}
 		saveFileToDB(fileToSave);
+		resendRequest(HttpMethod.POST, "/savedFile", filePath);
 		return ResponseEntity.ok("OK");
 	}
 
@@ -156,7 +185,28 @@ public class FilesController {
 		file.removeServer(server);
 		file = filesRepository.save(file);
 		server = serversRepository.save(server);
+		resendRequest(HttpMethod.DELETE, "/removeServerFromFile", removeInfo);
 		return ResponseEntity.ok("Removed " + serverid + " from " + filePath);
+	}
+
+	public void resendRequest(HttpMethod method, String urlPath, Object body) {
+		if (!zookeeperConfig.isLeader()) {
+			return;
+		}
+		List<Participant> participants = zookeeperConfig.getParticipants();
+		for (Participant participant: participants) {
+			String baseURL = participant.getId();
+			if (baseURL.equals(zookeeperConfig.leaderId())) {
+				continue;
+			}
+			String url = baseURL + urlPath;
+			System.out.println(baseURL);
+			RequestEntity<Object> request = RequestEntity.method(method, url).body(body);
+			log.info("Sending request " + request.toString() + " to server " + url);
+			RestTemplate template = new RestTemplate();
+			ResponseEntity<String> response = template.exchange(request, String.class);
+
+		}
 	}
 
 }
